+++ core/src/main/java/eu/kennytv/maintenance/core/util/DummySenderInfo.java
@@ -46,11 +46,6 @@ public final class DummySenderInfo implements SenderInfo {
     }
 
     @Override
-    @Deprecated
-    public void sendMessage(final String message) {
-    }
-
-    @Override
     public void send(final Component component) {
     }