+++ core/src/main/java/eu/kennytv/maintenance/core/util/SenderInfo.java
@@ -17,6 +17,8 @@
  */
 package eu.kennytv.maintenance.core.util;
 
+import eu.kennytv.maintenance.api.MaintenanceProvider;
+import eu.kennytv.maintenance.core.MaintenancePlugin;
 import eu.kennytv.maintenance.lib.kyori.adventure.text.Component;
 import eu.kennytv.maintenance.lib.kyori.adventure.text.minimessage.MiniMessage;
 import java.util.UUID;
@@ -33,13 +35,15 @@ public interface SenderInfo {
         return permission == null || hasPermission("maintenance.admin") || hasPermission("maintenance." + permission);
     }
 
-    @Deprecated
-    void sendMessage(String message);
-
     void send(Component component);
 
     default void sendRich(final String message) {
         send(MiniMessage.miniMessage().deserialize(message));
+    }
+
+    default void sendPrefixedRich(final String message) {
+        final MaintenancePlugin maintenance = (MaintenancePlugin) MaintenanceProvider.get();
+        send(maintenance.prefix().append(MiniMessage.miniMessage().deserialize(message)));
     }
 
     boolean isPlayer();